modified:   src/main/java/ifsp/edu/br/task_list/controller/ProjetoController.java 	modified:   src/main/java/ifsp/edu/br/task_list/model/Usuario.java 	modified:   src/main/java/ifsp/edu/br/task_list/repository/ProjetoRepository.java 	modified:   src/main/java/ifsp/edu/br/task_list/service/ProjetoService.java 	modified:   src/main/java/ifsp/edu/br/task_list/service/UsuarioService.java 	modified:   src/main/resources/data.sql

+++ src/main/java/ifsp/edu/br/task_list/repository/ProjetoRepository.java
@@ -9,9 +9,5 @@ import java.util.List;
 
 @Repository
 public interface ProjetoRepository extends JpaRepository<Projeto, Long> {
-
-    List<Projeto> findByUsuarioIdUsuario(Long idUsuario);
-
+    List<Projeto> findByUsuarioIdUsuario(Integer idUsuario);
 }
-
-
